ris: refactor ri basis parser
