+++ modules/ris/include/motis/ris/ribasis/ribasis_formation_parser.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace motis::ris::ribasis::formation {
+
+//
+
+}  // namespace motis::ris::ribasis::formation
