+++ modules/ris/src/ribasis/ribasis_formation_parser.cc
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace motis::ris::ribasis::formation {
+
+//
+
+}  // namespace motis::ris::ribasis::formation
